[Profiler] Move Kineto activity generation into `collection.cpp` (#80796)
In order to fold kineto events into TorchTidy, the new collection machinery needs to handle moving events to and from kineto. This PR is one step in that transition. In order to limit the scope of changes, we keep metadata generation in `profiler_kineto.cpp` for now.

Differential Revision: [D37406955](https://our.internmc.facebook.com/intern/diff/D37406955/)
Pull Request resolved: https://github.com/pytorch/pytorch/pull/80796
Approved by: https://github.com/aaronenyeshi